+++ modules/sd_vae_approx.py
@@ -5,6 +5,8 @@ from modules import devices, paths, shared
 
 
 sd_vae_approx_model = None
+simple_weights = None
+simple_bias = None
 
 
 class VAEApprox(nn.Module):
@@ -32,38 +34,49 @@ class VAEApprox(nn.Module):
         return x
 
 
-def model():
+def nn_approximation(sample): # Approximate NN
     global sd_vae_approx_model # pylint: disable=global-statement
     if sd_vae_approx_model is None:
-        from modules.shared import log
         model_path = os.path.join(paths.models_path, "VAE-approx", "model.pt")
         sd_vae_approx_model = VAEApprox()
         if not os.path.exists(model_path):
             model_path = os.path.join(paths.script_path, "models", "VAE-approx", "model.pt")
-        sd_vae_approx_model.load_state_dict(torch.load(model_path, map_location='cpu' if devices.device.type != 'cuda' else None))
+        approx_weights = torch.load(model_path, map_location='cpu' if devices.device.type != 'cuda' else None)
+        sd_vae_approx_model.load_state_dict(approx_weights)
         sd_vae_approx_model.eval()
         sd_vae_approx_model.to(devices.device, devices.dtype)
-        log.info(f"Loaded VAE-approx: model={model_path}")
-    return sd_vae_approx_model
-
-
-def cheap_approximation(sample):
-    # https://discuss.huggingface.co/t/decoding-latents-to-rgb-without-upscaling/23204/2
-    if shared.sd_model_type == "sdxl":
-        weight = torch.tensor([
-            [0.4543,-0.2868, 0.1566,-0.4748],
-            [0.5008, 0.0952, 0.2155,-0.3268],
-            [0.5294, 0.1625,-0.0624,-0.3793]
-        ]).reshape(3, 4, 1, 1).to(sample.device)
-        bias = torch.tensor([0.1375, 0.0144, -0.0675]).to(sample.device)
-    else:
-        weight = torch.tensor([
-            [0.298, 0.187,-0.158,-0.184],
-            [0.207, 0.286, 0.189,-0.271],
-            [0.208, 0.173, 0.264,-0.473],
-        ]).reshape(3, 4, 1, 1).to(sample.device)
-        bias = None
+        shared.log.debug(f'Loaded VAE decode approximate: model="{model_path}"')
     try:
-        return nn.functional.conv2d(sample, weight, bias) # pylint: disable=not-callable
-    except Exception:
+        in_sample = sample.to(devices.device, devices.dtype).unsqueeze(0)
+        x_sample = sd_vae_approx_model(in_sample)
+        x_sample = x_sample[0]
+        return x_sample
+    except Exception as e:
+        shared.log.error(f'Decode approximate: {e}')
+        return sample
+
+
+def cheap_approximation(sample): # Approximate simple
+    # https://discuss.huggingface.co/t/decoding-latents-to-rgb-without-upscaling/23204/2
+    global simple_weights, simple_bias # pylint: disable=global-statement
+    if simple_weights is None or simple_bias is None:
+        if shared.sd_model_type == "sdxl":
+            simple_weights = torch.tensor([
+                [0.4543,-0.2868, 0.1566,-0.4748],
+                [0.5008, 0.0952, 0.2155,-0.3268],
+                [0.5294, 0.1625,-0.0624,-0.3793]
+            ]).reshape(3, 4, 1, 1).to(sample.device)
+            simple_bias = torch.tensor([0.1375, 0.0144, -0.0675]).to(sample.device)
+        else:
+            simple_weights = torch.tensor([
+                [0.298, 0.187,-0.158,-0.184],
+                [0.207, 0.286, 0.189,-0.271],
+                [0.208, 0.173, 0.264,-0.473],
+            ]).reshape(3, 4, 1, 1).to(sample.device)
+            simple_bias = None
+    try:
+        x_sample = nn.functional.conv2d(sample, simple_weights, simple_bias) # pylint: disable=not-callable
+        return x_sample
+    except Exception as e:
+        shared.log.error(f'Decode simple: {e}')
         return sample